Make DDL hook test predictable across PG versions
The `ddl_hook` test failed on PG12 because dropped objects in the DDL
hook had a different order compared to previous versions. This change
therefore adds a sort of the dropped objects list before processing so
that the order objects when printed in the test is predictable.

The ordering happens on object type, which assumes the order of
objects with the same type is predictable. This seems sufficient for
the test to pass.