Improve target list handling for INSERTs
`INSERT`s on a hypertable requires wrapping the top-level
`ModifyTable` plan node with a `CustomScan` node (`HypertableInsert`)
that setups the tuple routing that occurs during execution. However,
using `CustomScan` nodes at the top-level requires implementing a
number of work-arounds for assumptions made in the planner on how
top-level nodes deal with targetlists.

The existing way this was handled wasn't robust enough, and problems
occur when enabling JIT compilation for expression evaluation. This
change improves the handling to work also for JIT.

Intuitively, our `HypertableInsert` node should adopt the target list
of the `ModifyTable` subplan without further projection. For a
`CustomScan` this means setting the "input" `custom_scan_tlist` to the
`ModifyTable`'s target list and having an "output" `targetlist` that
references the `TupleDesc` that is created from the
`custom_scan_tlist` at execution time. Now, while this seems
straight-forward, there are several things with how `ModifyTable`
nodes are handled in the planner that complicates this:

- First, `ModifyTable` doesn't set a `targetlist` when the node is
  created. It is only set later in `set_plan_references` (`setrefs.c`)
  if there's a `RETURNING` clause. Thus, there's no `targetlist`
  available when the `HypertableInsert` plan is created.

- Second, top-level plan nodes, except for `ModifyTable` nodes, need
  to have a `targetlist` matching `root->processed_tlist`. This is
  asserted in `apply_tlist_labeling`, which is called in `create_plan`
  (`createplan.c`) immediately after the `HypertableInsert` plan node
  is created. `ModifyTable` is exempted because it doesn't always have
  a `targetlist` that matches `processed_tlist`. So, even if we had
  access to `ModifyTable`'s targetlist at plan creation we wouldn't be
  able to use it since the `HypertableInsert` is a `CustomScan` and
  thus not exempted.

- Third, a `CustomScan`'s `targetlist` should reference the attributes
  of the `TupleDesc` that gets created from the `custom_scan_tlist` at
  the start of execution. This means we need to make the `targetlist`
  into all `Var`s with attribute numbers that correspond to the
  `TupleDesc` instead of result relation in the `ModifyTable`.

To get around these issues, we set the `CustomScan`'s `targetlist` to
`root->processed_tlist` when we create the plan node, and at the end
of planning when the `ModifyTable`'s `targetlist` is set, we go back
and fix up the `CustomScan`'s `targetlist`.