Add debug printout for malformed telemetry
If the telemetry response is malformed, strange errors will be
generated in the log because the use of `DirectFunctionCall2` expect
the result of function calls to not be NULL and will throw an error if
it is not.

By printing the response in the log we can debug what went wrong.