Update append-12 output after rebase
Update the PG12-specific append test. Earlier, the test produced
a different chunk order in plans across PostgreSQL versions because
the ordering of the inserted data was not deterministic.
The ordering was made deterministic and this change updates the PG12
output after these changes, now making the plan output consistent
with other PostgreSQL versions.